# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Maryam Abacha American University of Nigeria - University Management Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
        - name: semester
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a single course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List enrollments (student sees own, admin/lecturer sees all)
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: courseId
          in: query
          schema:
            type: integer
        - name: semester
          in: query
          schema:
            type: string
        - name: academicYear
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: enrollCourse
      tags: [enrollments]
      summary: Register/enroll in a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollCourseBody"
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
        "409":
          description: Already enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /enrollments/{id}:
    delete:
      operationId: dropCourse
      tags: [enrollments]
      summary: Drop a course enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dropped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /results:
    get:
      operationId: listResults
      tags: [results]
      summary: List results
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: courseId
          in: query
          schema:
            type: integer
        - name: semester
          in: query
          schema:
            type: string
        - name: academicYear
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
    post:
      operationId: submitResult
      tags: [results]
      summary: Submit or update a result (lecturer/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitResultBody"
      responses:
        "201":
          description: Result submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /results/{id}:
    put:
      operationId: updateResult
      tags: [results]
      summary: Update a result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResultBody"
      responses:
        "200":
          description: Updated result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        matricNumber:
          type: string
        department:
          type: string
        faculty:
          type: string
        level:
          type: string
        cgpa:
          type: number
          nullable: true
        enrollmentYear:
          type: string
        createdAt:
          type: string
      required: [id, userId, name, email, matricNumber, department, faculty, level, enrollmentYear, createdAt]
...
    Course:
      type: object
      properties:
        id:
          type: integer
        courseCode:
          type: string
        title:
          type: string
        unit:
          type: integer
        department:
          type: string
        faculty:
          type: string
        level:
          type: string
        semester:
          type: string
          enum: [first, second]
        description:
          type: string
          nullable: true
        lecturerId:
          type: integer
          nullable: true
        lecturerName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, courseCode, title, unit, department, faculty, level, semester, createdAt]
    CreateCourseBody:
      type: object
      properties:
        courseCode:
          type: string
        title:
          type: string
        unit:
          type: integer
        department:
          type: string
        faculty:
          type: string
        level:
          type: string
        semester:
          type: string
          enum: [first, second]
        description:
          type: string
        lecturerId:
          type: integer
      required: [courseCode, title, unit, department, faculty, level, semester]
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        courseId:
          type: integer
        semester:
          type: string
        academicYear:
          type: string
        status:
          type: string
          enum: [active, dropped, completed]
        course:
          $ref: "#/components/schemas/Course"
          nullable: true
        student:
          $ref: "#/components/schemas/StudentProfile"
          nullable: true
        enrolledAt:
          type: string
      required: [id, studentId, courseId, semester, academicYear, status, enrolledAt]
    EnrollCourseBody:
      type: object
      properties:
        courseId:
          type: integer
        semester:
          type: string
        academicYear:
          type: string
      required: [courseId, semester, academicYear]
    Result:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        courseId:
          type: integer
        semester:
          type: string
        academicYear:
          type: string
        caScore:
          type: number
          nullable: true
        examScore:
          type: number
          nullable: true
        totalScore:
          type: number
          nullable: true
        grade:
          type: string
          nullable: true
        gradePoint:
          type: number
          nullable: true
        course:
          $ref: "#/components/schemas/Course"
          nullable: true
        student:
          $ref: "#/components/schemas/StudentProfile"
          nullable: true
        createdAt:
          type: string
      required: [id, studentId, courseId, semester, academicYear, createdAt]
    SubmitResultBody:
      type: object
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
        semester:
          type: string
        academicYear:
          type: string
        caScore:
          type: number
        examScore:
          type: number
      required: [studentId, courseId, semester, academicYear]
    UpdateResultBody:
      type: object
      properties:
        caScore:
          type: number
        examScore:
          type: number
      required: []
...